finish detailpage

+++ lib/models/space.dart
@@ -1,8 +1,9 @@
+
 class Space {
   int id;
   String name;
   String imageUrl;
-  double price;
+  int price;
   String city;
   String country;
   String status;
